TGBOT-78 Rename class
